refactor: 패키지 구조 및 클래스 네이밍 리팩토링
