+++ src/main/java/com/izza/search/persistent/query/MapSearchQuery.java
@@ -1,4 +1,4 @@
-package com.izza.search.persistent;
+package com.izza.search.persistent.query;
 
 import com.izza.search.domain.ZoomLevel;
 import com.izza.search.vo.Point;
@@ -6,7 +6,7 @@ import com.izza.search.vo.Point;
 /**
  * 행정구역 검색 요청 DTO (Persistent 계층용)
  */
-public record AreaSearchQuery(
+public record MapSearchQuery(
         ZoomLevel zoomLevel,
         Point southWest,
         Point northEast
